Implement only execute list of task
* Update versions
* Update run script
* Update docs
