+++ .env
@@ -1,2 +1,1 @@
 AIRFLOW_UID=501
-AIRFLOW_GID=0
