+++ Dockerfile
@@ -1,3 +1,3 @@
 FROM apache/airflow:latest
-COPY requirements.txt .
-RUN pip install -r requirements.txt
+# ADD requirements.txt .
+RUN pip install apache-airflow==${AIRFLOW_VERSION} loguru hydra-core
