+++ colima.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-# This is a helper script to manage running Airflow locally.
+# Here is a helper script for running Airflow locally
 
 # Fonts utils
 RED_COLOR="\033[0;31m"
@@ -10,15 +10,14 @@ BOLD=$(tput bold)
 NC=$(tput sgr0)
 
 start_airflow() {
-    colima stop
-
     colima start \
         --cpu 4 \
         --memory 8 \
         --mount-type sshfs \
         --mount $(PWD)/dags:w \
         --mount $(PWD)/logs:w \
-        --mount $(PWD)/plugins:w
+        --mount $(PWD)/plugins:w \
+        --mount $(PWD)/config:w
 
     docker-compose up -d
 }
@@ -28,15 +27,16 @@ restart_airflow() {
     docker-compose up -d
 }
 
-deleter_airflow() {
+delete_airflow() {
     docker-compose down --volumes --rmi local
+    colima stop
 }
 
 main() {
     echo -e "${GREEN_COLOR}"
     echo -e "###############################################################################"
     echo -e "################################ Local Aiflow #################################"
-    echo -e "######### This is a helper script to manage running Airflow locally. ##########"
+    echo -e "############# Here is a helper script for running Airflow locally #############"
     echo -e "###############################################################################"
     echo -e "${NC}"
     echo -e "${GREEN_COLOR} 1. Start Airflow"
@@ -48,7 +48,7 @@ main() {
     elif [ $option = "2" ]; then
         restart_airflow
     elif [ $option = "3" ]; then
-        deleter_airflow
+        delete_airflow
     else
         echo -e "${RED_COLOR}Invalid option. Exiting...${NC}"
         exit 1
